moved pricing sections into components
